electron/python/requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# OCR and Computer Vision
rapidocr-onnxruntime==1.3.21
opencv-python==4.8.1.78

# Image Processing
pillow==10.0.1
pdf2image==1.16.3

# Data Processing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4

# Excel/Spreadsheet Support
openpyxl==3.1.2
pyexcel-ods3==0.6.1
odfpy==1.4.1

# Visualization (optional)
matplotlib==3.7.2

# System Dependencies (already in Python standard library)
# logging - built-in
# tempfile - built-in
# base64 - built-in
# os - built-in
# sys - built-in
# io - built-in
# collections - built-in